+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [Add a checkbox to intake form to prevent email to stakeholder (only visible to staff) #99](https://github.com/farmier/farm_rcd/issues/99)
+
 ### Fixed
 
 - [Fix access to View of plans by farm #120](https://github.com/farmier/farm_rcd/issues/120)